chore: move FormatForm to explorer components
