+++ packages/frontend/src/components/Explorer/CustomMetricModal/index.tsx
@@ -25,11 +25,11 @@ import { useForm } from '@mantine/form';
 import { useCallback, useEffect, useMemo, useState } from 'react';
 import { ValueOf } from 'type-fest';
 import { v4 as uuidv4 } from 'uuid';
-import { FormatForm } from '../../../features/tableCalculation/components/FormatForm';
 import useToaster from '../../../hooks/toaster/useToaster';
 import { useExplore } from '../../../hooks/useExplore';
 import { useExplorerContext } from '../../../providers/ExplorerProvider';
 import { FiltersProvider } from '../../common/Filters/FiltersProvider';
+import { FormatForm } from '../FormatForm';
 import { FilterForm, MetricFilterRuleWithFieldId } from './FilterForm';
 import { useDataForFiltersProvider } from './hooks/useDataForFiltersProvider';
 import {
